gpu_worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pillow==10.4.0
numpy==1.26.4
insightface==0.7.3
onnxruntime-gpu==1.16.3
opencv-python==4.10.0.84
aiofiles==23.2.1
